fix: Serial2.read() must receive int8_t
fix for 3d376a8ec7e12be7357d27e5f7536af2f40c8e00

+++ mainprogram/Program/Program.ino
@@ -172,7 +172,9 @@ void loop() {
     // get gyrodata
     if (Serial2.available() > 0) {
         while (Serial2.available() != 0) {  //  Gyro の方位データを gyro に取り込む
-            gyro_o = Serial2.read();
+            // Serial2の送信側がint8_tで送ってるので、intで受け取ると負の数が期待通り受け取れない。
+            // そのため、int8_tにキャストする必要がある。
+            gyro_o = (int8_t)Serial2.read();
         }
     }
